Move ui components to ui & stateful components to components folder
