+++ src/app/index.tsx
@@ -1,73 +1,18 @@
-import { useCallback, useEffect } from 'react';
-import { UIProvider } from 'components/UIProvider';
-import { Box } from 'components/Box';
-import { Chip } from 'components/Chip';
-import { PomodoroIntervals } from 'types';
-import { useAppDispatch, useAppSelector } from 'store';
-import {
-  incrementPomodoroCount,
-  resetPomodoroCount,
-  selectCurrentInterval,
-  selectPomodoroCount,
-  selectPomodorosInSession,
-  setCurrentInterval,
-} from 'store/intervals';
-import { TimerProvider } from 'components/Timer';
-import { useThemedFavicon } from 'hooks';
+import { UIProvider } from 'ui/UIProvider';
+import { Box } from 'ui/Box';
+import { TimerProvider } from 'ui/Timer';
+import { useThemedFavicon, useIntervalSequence } from 'hooks';
 import { ControlPanel } from 'components/ControlPanel';
-
-const chipLabels: Record<PomodoroIntervals, string> = {
-  pomodoro: 'Focus',
-  shortBreak: 'Short Break',
-  longBreak: 'Long Break',
-};
+import { IntervalChip } from 'components/IntervalChip';
 
 export const App = () => {
-  const dispatch = useAppDispatch();
-
-  const pomodorosInSession = useAppSelector(selectPomodorosInSession);
-  const currentInterval = useAppSelector(selectCurrentInterval);
-  const pomodoroCount = useAppSelector(selectPomodoroCount);
-
-  const label = chipLabels[currentInterval];
-
-  const getNextInterval = useCallback((): PomodoroIntervals => {
-    if (currentInterval === 'shortBreak' || currentInterval === 'longBreak') {
-      return 'pomodoro';
-    }
-
-    if (
-      currentInterval === 'pomodoro' &&
-      pomodoroCount === pomodorosInSession - 1
-    ) {
-      return 'longBreak';
-    }
-
-    return 'shortBreak';
-  }, [currentInterval, pomodoroCount, pomodorosInSession]);
-
-  const handleTimerFinish = useCallback(() => {
-    const nextInterval = getNextInterval();
-
-    if (nextInterval === 'longBreak') {
-      dispatch(resetPomodoroCount());
-    } else if (nextInterval !== 'pomodoro') {
-      dispatch(incrementPomodoroCount());
-    }
-
-    dispatch(setCurrentInterval(nextInterval));
-  }, [dispatch, getNextInterval]);
-
-  useEffect(() => {
-    dispatch(setCurrentInterval('pomodoro'));
-    dispatch(resetPomodoroCount());
-  }, [dispatch, pomodorosInSession]);
+  const { onTimerFinish } = useIntervalSequence();
 
   useThemedFavicon();
 
   return (
     <UIProvider>
-      <TimerProvider onTimerFinish={handleTimerFinish}>
+      <TimerProvider onTimerFinish={onTimerFinish}>
         <Box
           display="flex"
           alignItems="center"
@@ -75,9 +20,7 @@ export const App = () => {
           fullWidth
           mt={8}
         >
-          <Chip icon={currentInterval === 'pomodoro' ? 'focus' : 'break'}>
-            {label}
-          </Chip>
+          <IntervalChip />
 
           <Box mt={8}>
             <ControlPanel />
