--- src/components/Modal/index.ts
@@ -1,2 +1,0 @@
-export * from './Modal';
-export * from './SettingsModal';
