+++ src/hooks/index.ts
@@ -1,2 +1,3 @@
+export * from './useIntervalSequence';
 export * from './useSettings';
 export * from './useThemedFavicon';
