+++ src/ui/Icon/static/Break.tsx
