+++ src/ui/Icon/static/Close.tsx
