+++ src/ui/Icon/static/Dots.tsx
