+++ src/ui/Icon/static/Focus.tsx
