+++ src/ui/Icon/static/Forward.tsx
