+++ src/ui/Icon/static/Pause.tsx
