+++ src/ui/Icon/static/Play.tsx
