+++ src/ui/Switch/Switch.styles.ts
