+++ src/ui/Switch/index.tsx
@@ -1,5 +1,5 @@
 import { ChangeEvent, useCallback, useState } from 'react';
-import { TextFieldProps } from 'components/TextField';
+import { TextFieldProps } from 'ui/TextField';
 import { styles } from './Switch.styles';
 
 export const Switch = ({
